fix(appeals): appeal withdrawn email for lpa incorrect (a2-2739)
