+++ appeals/api/src/server/endpoints/withdrawal/withdrawal.service.js
@@ -31,11 +31,15 @@ export const publishWithdrawal = async (
 	const recipientEmail = appeal.agent?.email || appeal.appellant?.email;
 	const lpaEmail = appeal.lpa?.email || '';
 
+	const eventType = appeal.appealStatus[0].status === APPEAL_CASE_STATUS.EVENT ? 'site visit' : '';
+
 	const personalisation = {
 		appeal_reference_number: appeal.reference,
 		lpa_reference: appeal.applicationReference || '',
 		site_address: siteAddress,
-		withdrawal_date: formatDate(new Date(withdrawalRequestDate || ''), false)
+		withdrawal_date: formatDate(new Date(withdrawalRequestDate || ''), false),
+		event_type: eventType,
+		event_set: !!eventType
 	};
 
 	if (recipientEmail) {
